feat: create LoginForm component for user authentication
